data_engineering_pipeline/data_engineering_pipeline/requirements.txt
beautifulsoup4
pymongo
requests
python-dotenv
tqdm
pytest